Feat : 상품 등록 시 카테고리 전체 조회 로직 추가 완성
